+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 dd mm yyyy
 
+- Fix: Search for character names containing a slash was broken.
 - Improved texts.
 
 ## 2.2.0